refactor peer_conductor hpp => cpp
