+++ Plugins/NativeServerPlugin/StreamingNativeServerPlugin.vcxproj
@@ -172,7 +172,6 @@
     <Lib />
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="inc\peer_conductor.hpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">pch.h</PrecompiledHeaderFile>
@@ -183,11 +182,14 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Release|x64'">pch.h</PrecompiledHeaderFile>
     </ClCompile>
+    <ClCompile Include="peer_conductor.cpp" />
     <ClCompile Include="src\buffer_capturer.cpp" />
     <ClCompile Include="src\conductor.cpp" />
     <ClCompile Include="src\defaults.cpp" />
     <ClCompile Include="src\directx_buffer_capturer.cpp" />
+    <ClCompile Include="src\directx_peer_conductor.cpp" />
     <ClCompile Include="src\input_data_channel_observer.cpp" />
+    <ClCompile Include="src\peer_view.cpp" />
     <ClCompile Include="src\render_service.cpp" />
     <ClCompile Include="src\service_base.cpp" />
   </ItemGroup>
@@ -196,8 +198,11 @@
     <ClInclude Include="inc\conductor.h" />
     <ClInclude Include="inc\defaults.h" />
     <ClInclude Include="inc\directx_buffer_capturer.h" />
+    <ClInclude Include="inc\directx_peer_conductor.h" />
     <ClInclude Include="inc\input_data_channel_observer.h" />
     <ClInclude Include="inc\multi_peer_conductor.hpp" />
+    <ClInclude Include="inc\peer_conductor.h" />
+    <ClInclude Include="inc\peer_view.h" />
     <ClInclude Include="inc\plugindefs.h" />
     <ClInclude Include="inc\flagdefs.h" />
     <ClInclude Include="inc\macros.h" />
