+++ Plugins/NativeServerPlugin/StreamingNativeServerPlugin.vcxproj.filters
@@ -55,8 +55,14 @@
     <ClCompile Include="src\directx_buffer_capturer.cpp">
       <Filter>Source\StreamingToolkit</Filter>
     </ClCompile>
-    <ClCompile Include="inc\peer_conductor.hpp">
+    <ClCompile Include="peer_conductor.cpp">
       <Filter>Source\webrtc</Filter>
+    </ClCompile>
+    <ClCompile Include="src\peer_view.cpp">
+      <Filter>Source\StreamingToolkit</Filter>
+    </ClCompile>
+    <ClCompile Include="src\directx_peer_conductor.cpp">
+      <Filter>Source\StreamingToolkit</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -96,6 +102,15 @@
     </ClInclude>
     <ClInclude Include="inc\multi_peer_conductor.hpp">
       <Filter>Source\webrtc</Filter>
+    </ClInclude>
+    <ClInclude Include="inc\peer_conductor.h">
+      <Filter>Headers\webrtc</Filter>
+    </ClInclude>
+    <ClInclude Include="inc\peer_view.h">
+      <Filter>Headers\StreamingToolkit</Filter>
+    </ClInclude>
+    <ClInclude Include="inc\directx_peer_conductor.h">
+      <Filter>Headers\StreamingToolkit</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
